Reorganize macros
